refactor(common): bettafish_analyzer 5분할 파사드 유지 (P2-B)
1,625줄 단일 파일을 책임 단위로 4개 leaf/중간 모듈 + 파사드로 분리.
행위 보존 리팩토링으로 공개 API·import 경로 불변.

- bettafish_models.py: ReportChapter, AnalysisReport + 헬퍼 4종 (leaf)
- bettafish_synthesis.py: ForumSynthesis (models 의존)
- bettafish_insight.py: InsightForge (models 의존)
- bettafish_perspectives.py: Data/Sentiment/Macro Perspective (models 의존)
- bettafish_analyzer.py: BettaFishAnalyzer + 편의 함수 + 상단 재-export

파사드가 상단에서 4개 모듈을 재-export(# noqa: F401)하여
`from common.bettafish_analyzer import ...` 소비자(collect_market_indicators,
collect_coinmarketcap)와 테스트 import 경로가 변경 없이 동작.
test_bettafish_analyzer.py / test_import_compatibility.py diff 0.
docs/component-counts.md 공통 모듈 수 57→61 동기화.

게이트: ruff check/format 통과, basedpyright 0 errors,
pytest 159 passed(대상 서브셋), 전체 스위트 4689 passed.
